backend/requirements.txt
flask==3.1.0
anthropic==0.49.0
requests==2.32.3
python-dotenv==1.1.0
pytest==8.3.5
pytest-cov==6.1.0
gunicorn==23.0.0 
